Account: switch error handling to toasts
