+++ components/brave_account/resources/brave_account_page.html
@@ -23,6 +23,7 @@
       <script src="//resources/js/ios/mojo_api.js"></script>
       <script src="//resources/js/ios/web_ui.js"></script>
     </if>
+    <leo-alertcenter></leo-alertcenter>
     <brave-account-dialogs></brave-account-dialogs>
   </body>
 </html>
