+++ components/brave_account/resources/brave_account_sign_in_dialog.ts
@@ -11,8 +11,8 @@ import {
 } from './brave_account_browser_proxy.js'
 import { getCss } from './brave_account_sign_in_dialog.css.js'
 import { getHtml } from './brave_account_sign_in_dialog.html.js'
-import { Error } from './brave_account_common.js'
 import { LoginError, LoginErrorCode } from './brave_account.mojom-webui.js'
+import { showError } from './brave_account_common.js'
 
 // @ts-expect-error
 import { Login } from 'chrome://resources/brave/opaque_ke.bundle.js'
@@ -81,10 +81,7 @@ export class BraveAccountSignInDialogElement extends CrLitElement {
         details = { netErrorOrHttpStatus: null, errorCode: null }
       }
 
-      this.fire('error-occurred', {
-        flow: 'login',
-        details,
-      } satisfies Extract<Error, { flow: 'login' }>)
+      showError({ kind: 'login', details })
     }
   }
